+++ crates/dslab-faas-extra/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 [dependencies]
 arima = "0.2"
 csv = "1.1"
+dslab-faas = { path = "../dslab-faas" }
 num = "0.2.0"
 rand = "0.8"
-dslab-faas = { path = "../dslab-faas" }
+rand_pcg = "0.3.1"
