+++ crates/dslab-faas-extra/src/lib.rs
@@ -1,3 +1,5 @@
 pub(crate) mod arima_extra;
+pub mod azure_experiment;
 pub mod hybrid_histogram;
 pub mod opendc_trace;
+pub mod simple_schedulers;
